backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
cached-property==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.4
django-appconf==1.0.3
django-cors-headers==3.4.0
django-heroku==0.3.1
django-import-export==2.0.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-solo==1.1.3
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
isodate==0.6.0
jdcal==1.4.1
jsonfield==3.1.0
lxml==4.5.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
tablib==1.1.0
urllib3==1.25.9
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
zeep==3.4.0
